requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio
torchmetrics
tokenizers
editdistance
transformers
soundfile
numpy==1.26.4
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
librosa
sentencepiece

black
isort
pre-commit
wget
